refactor: dedupe test fakes and unify sent-date resolution
- hoist the archiver's COM fakes and the migration property-read fake
  into tests/conftest.py so every test module stops re-declaring them
- replace the flag_status/message_id migration test triples with one
  test parametrized over models._COLUMNS, so the next column added
  there is covered for free instead of copied a third time
- have archiver._get_sent_date_prefix format outlook.client's
  _sent_datetime instead of re-implementing the SentOn/ReceivedTime
  fallback with its own, looser exception handling -- the two could
  previously disagree on a mail's date

Closes #63
