+++ tests/test_archiver_attachment_dedupe.py
@@ -25,40 +25,7 @@ from email_archiver.archiver.archiver import (
     EmailArchiver,
 )
 
-
-class _FakeAttachment:
-    """A real (non-inline) attachment: no MAPI properties, so the archiver's
-    inline detection falls through both branches and saves it."""
-
-    def __init__(self, filename):
-        self.FileName = filename
-
-    @property
-    def PropertyAccessor(self):
-        raise AttributeError("no PropertyAccessor on this fake")
-
-    def SaveAsFile(self, path):  # noqa: N802 - COM-shaped API
-        with open(path, "w", encoding="utf-8") as fh:
-            fh.write("att")
-
-
-class _FakeAttachments:
-    def __init__(self, items=()):
-        self._items = list(items)
-
-    def __iter__(self):
-        return iter(self._items)
-
-
-class _FakeMailItem:
-    def __init__(self, attachments=()):
-        self.Attachments = _FakeAttachments(
-            _FakeAttachment(n) for n in attachments
-        )
-
-    def SaveAs(self, path, fmt):  # noqa: N802 - COM-shaped API
-        with open(path, "w", encoding="utf-8") as fh:
-            fh.write("msg")
+from .conftest import _FakeMailItem
 
 
 def _run_bounded(target, *, timeout=10.0):
